+++ src/views/Home.vue
@@ -53,6 +53,9 @@
 <style>
     .clickable {
         cursor:pointer;
+        -webkit-user-select: none; /* Chrome/Safari */
+        -moz-user-select: none; /* Firefox */
+        -ms-user-select: none; /* IE10+ */
     }
 
     .odd-row {